@@ -11,8 +11,10 @@
   },
   "dependencies": {
     "async": "^2.1.2",
+    "autoprefixer": "^7.1.2",
     "babel-core": "^6.24.1",
-    "babel-loader": "^6.0.0",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-remove-graphql-queries": "1.4.1",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-object-assign": "^6.8.0",
     "babel-preset-env": "^1.6.0",
@@ -27,20 +29,20 @@
     "bundle-loader": "^0.5.4",
     "chalk": "^1.1.3",
     "chokidar": "^1.7.0",
-    "chunk-manifest-webpack-plugin": "0.1.0",
+    "chunk-manifest-webpack-plugin": "^1.1.2",
     "commander": "^2.9.0",
     "common-tags": "^1.4.0",
     "convert-hrtime": "^2.0.0",
     "copyfiles": "^1.2.0",
-    "css-loader": "^0.26.1",
+    "css-loader": "^0.28.4",
     "debug": "^2.6.0",
     "detect-port": "^1.2.1",
     "dotenv": "^4.0.0",
     "eventemitter2": "^4.1.0",
     "express": "^4.14.0",
     "express-graphql": "^0.6.6",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
+    "extract-text-webpack-plugin": "^2.0.0",
+    "file-loader": "^1.0.0",
     "flat": "^2.0.1",
     "friendly-errors-webpack-plugin": "^1.6.1",
     "front-matter": "^2.1.0",
@@ -81,9 +83,10 @@
     "parse-filepath": "^1.0.1",
     "path-exists": "^3.0.0",
     "postcss-browser-reporter": "^0.5.0",
-    "postcss-cssnext": "^2.8.0",
-    "postcss-import": "^8.2.0",
-    "postcss-loader": "^0.13.0",
+    "postcss-cssnext": "^3.0.2",
+    "postcss-flexbugs-fixes": "^3.0.0",
+    "postcss-import": "^10.0.0",
+    "postcss-loader": "^2.0.6",
     "postcss-reporter": "^1.4.1",
     "raw-loader": "^0.5.1",
     "react": "^15.6.0",
@@ -101,17 +104,17 @@
     "socket.io": "^2.0.3",
     "static-site-generator-webpack-plugin": "^3.1.0",
     "string-similarity": "^1.2.0",
-    "style-loader": "^0.13.0",
+    "style-loader": "^0.18.2",
     "tracer": "^0.8.7",
     "type-of": "^2.0.1",
     "underscore.string": "^3.3.4",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.3",
-    "webpack-configurator": "^0.3.0",
-    "webpack-dev-middleware": "^1.8.4",
-    "webpack-dev-server": "^1.16.1",
-    "webpack-hot-middleware": "^2.13.2",
+    "url-loader": "^0.5.9",
+    "webpack": "^2.0.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-dev-server": "^2.6.1",
+    "webpack-hot-middleware": "^2.18.2",
     "webpack-md5-hash": "0.0.5",
+    "webpack-merge": "^4.1.0",
     "webpack-stats-plugin": "^0.1.4",
     "webpack-validator": "^2.2.7",
     "yaml-loader": "^0.4.0",
@@ -152,6 +155,6 @@
     "build:cli": "babel src/gatsby-cli.js --out-file dist/gatsby-cli.js --presets es2015",
     "clean-test-bundles": "find test/ -type f -name bundle.js* -exec rm -rf {} +",
     "test-coverage": "node_modules/.bin/nyc --reporter=lcov --reporter=text npm test",
-    "watch": "rimraf dist && mkdir dist && npm run build:cli && npm run build:src -- --watch && npm run build:internal-plugins && npm run build:rawfiles"
+    "watch": "rimraf dist && mkdir dist && npm run build:cli && npm run build:internal-plugins && npm run build:rawfiles && npm run build:src -- --watch"
   }
 }
