@@ -1,3 +1,5 @@
 import mitt from "mitt"
+
 const emitter = mitt()
-module.exports = emitter
+export default emitter
+
