@@ -1,64 +1,60 @@
-import ExtractTextPlugin from "extract-text-webpack-plugin"
 
-exports.modifyWebpackConfig = ({ config, stage }) => {
-  const cssModulesConf = `css?modules&minimize&importLoaders=1`
-  const cssModulesConfDev = `${cssModulesConf}&sourceMap&localIdentName=[name]---[local]---[hash:base64:5]`
+exports.modifyWebpackConfig = ({ boundActionCreators, stage, rules, plugins, loaders }, { postCssPlugins, ...sassOptions }) => {
+  const { setWebpackConfig } = boundActionCreators
+  const PRODUCTION = stage !== `develop`
+
+  const sassLoader = {
+    loader: require.resolve(`sass-loader`),
+    options: {
+      sourceMap: !PRODUCTION,
+      ...sassOptions,
+    },
+  }
+
+  const sassRule = {
+    test: /\.s(a|c)ss$/,
+    exclude: /\.module\.s(a|c)ss$/,
+    use: plugins.extractText.extract({
+      fallback: loaders.style,
+      use: [loaders.css({ importLoaders: 1 }), loaders.postcss({ plugins: postCssPlugins }), sassLoader],
+    }),
+  }
+  const sassRuleModules = {
+    test: /\.module\.s(a|c)ss$/,
+    use: plugins.extractText.extract({
+      fallback: loaders.style,
+      use: [
+        loaders.css({ modules: true, importLoaders: 1 }),
+        loaders.postcss({ plugins: postCssPlugins }),
+        sassLoader,
+      ],
+    }),
+  }
+
+  let configRules = []
 
   switch (stage) {
-    case `develop`: {
-      config.loader(`sass`, {
-        test: /\.s(a|c)ss$/,
-        exclude: /\.module\.s(a|c)ss$/,
-        loaders: [`style`, `css`, `sass`],
-      })
+    case `develop`:
+    case `build-css`:
+    case `build-javascript`:
+      configRules = configRules.concat([sassRule, sassRuleModules])
+      break
 
-      config.loader(`sassModules`, {
-        test: /\.module\.s(a|c)ss$/,
-        loaders: [`style`, cssModulesConfDev, `sass`],
-      })
-      return config
-    }
-    case `build-css`: {
-      config.loader(`sass`, {
-        test: /\.s(a|c)ss$/,
-        exclude: /\.module\.s(a|c)ss$/,
-        loader: ExtractTextPlugin.extract([`css?minimize`, `sass`]),
-      })
-
-      config.loader(`sassModules`, {
-        test: /\.module\.s(a|c)ss$/,
-        loader: ExtractTextPlugin.extract(`style`, [cssModulesConf, `sass`]),
-      })
-      return config
-    }
-    case `build-html`: {
-      config.loader(`sass`, {
-        test: /\.s(a|c)ss$/,
-        exclude: /\.module\.s(a|c)ss$/,
-        loader: `null`,
-      })
-
-      config.loader(`sassModules`, {
-        test: /\.module\.s(a|c)ss$/,
-        loader: ExtractTextPlugin.extract(`style`, [cssModulesConf, `sass`]),
-      })
-      return config
-    }
-    case `build-javascript`: {
-      config.loader(`sass`, {
-        test: /\.s(a|c)ss$/,
-        exclude: /\.module\.s(a|c)ss$/,
-        loader: `null`,
-      })
-
-      config.loader(`sassModules`, {
-        test: /\.module\.s(a|c)ss$/,
-        loader: ExtractTextPlugin.extract(`style`, [cssModulesConf, `sass`]),
-      })
-      return config
-    }
-    default: {
-      return config
-    }
+    case `build-html`:
+    case `develop-html`:
+      configRules = configRules.concat([
+        {
+          ...sassRule,
+          use: loaders.null,
+        },
+        sassRuleModules,
+      ])
+      break
   }
+
+  setWebpackConfig({
+    module: {
+      rules: configRules,
+    },
+  })
 }
