@@ -18,104 +18,70 @@
  *   },
  * ],
  */
-const ExtractTextPlugin = require(`extract-text-webpack-plugin`)
+exports.modifyWebpackConfig = (
+  { boundActionCreators, stage, rules, plugins, loaders },
+  { postCssPlugins, ...stylusOptions }
+) => {
+  const { setWebpackConfig } = boundActionCreators
+  const PRODUCTION = stage !== `develop`
 
-exports.modifyWebpackConfig = ({ config, stage }, options = {}) => {
-  const cssModulesConfProd = `css?modules&minimize&importLoaders=1`
-  const cssModulesConfDev = `css?modules&importLoaders=1&localIdentName=[name]---[local]---[hash:base64:5]`
-
-  // Pass in stylus plugins regardless of stage.
-  if (Array.isArray(options.use)) {
-    config.merge(current => {
-      current.stylus = {
-        use: options.use,
-      }
-      return current
-    })
-  } else if (options.use) {
-    throw new Error(
-      `gatsby-plugin-stylus "use" option passed with ${options.use}. Pass an array of stylus plugins instead`
-    )
+  const stylusLoader = {
+    loader: require.resolve(`stylus-loader`),
+    options: {
+      sourceMap: !PRODUCTION,
+      ...stylusOptions,
+    },
   }
 
-  const stylusFiles = /\.styl$/
-  const stylusModulesFiles = /\.module\.styl$/
+  const stylusRule = {
+    test: /\.styl$/,
+    exclude: /\.module\.styl$/,
+    use: plugins.extractText.extract({
+      fallback: loaders.style,
+      use: [
+        loaders.css({ importLoaders: 1 }),
+        loaders.postcss({ plugins: postCssPlugins }),
+        stylusLoader,
+      ],
+    }),
+  }
+
+  const stylusRuleModules = {
+    test: /\.module\.styl$/,
+    use: plugins.extractText.extract({
+      fallback: loaders.style,
+      use: [
+        loaders.css({ modules: true, importLoaders: 1 }),
+        loaders.postcss({ plugins: postCssPlugins }),
+        stylusLoader,
+      ],
+    }),
+  }
+
+  let configRules = []
 
   switch (stage) {
-    case `develop`: {
-      config.loader(`stylus`, {
-        test: stylusFiles,
-        exclude: stylusModulesFiles,
-        loaders: [`style`, `css`, `postcss`, `stylus`],
-      })
-      config.loader(`stylusModules`, {
-        test: stylusModulesFiles,
-        loaders: [`style`, cssModulesConfDev, `postcss`, `stylus`],
-      })
-      return config
-    }
+    case `develop`:
+    case `build-css`:
+    case `build-javascript`:
+      configRules = configRules.concat([stylusRule, stylusRuleModules])
+      break
 
-    case `build-css`: {
-      config.loader(`stylus`, {
-        test: stylusFiles,
-        exclude: stylusModulesFiles,
-        loader: ExtractTextPlugin.extract(`style`, [
-          `css?minimize`,
-          `postcss`,
-          `stylus`,
-        ]),
-      })
-      config.loader(`stylusModules`, {
-        test: stylusModulesFiles,
-        loader: ExtractTextPlugin.extract(`style`, [
-          cssModulesConfProd,
-          `postcss`,
-          `stylus`,
-        ]),
-      })
-      return config
-    }
-
+    case `build-html`:
     case `develop-html`:
-    case `build-html`: {
-      const moduleLoader = ExtractTextPlugin.extract(`style`, [
-        cssModulesConfProd,
-        `postcss`,
-        `stylus`,
+      configRules = configRules.concat([
+        {
+          ...stylusRule,
+          use: loaders.null,
+        },
+        stylusRuleModules,
       ])
-
-      config.loader(`stylus`, {
-        test: stylusFiles,
-        exclude: stylusModulesFiles,
-        loader: `null`,
-      })
-      config.loader(`stylusModules`, {
-        test: stylusModulesFiles,
-        loader: moduleLoader,
-      })
-      return config
-    }
-
-    case `build-javascript`: {
-      const moduleLoader = ExtractTextPlugin.extract(`style`, [
-        cssModulesConfProd,
-        `stylus`,
-      ])
-      config.loader(`stylus`, {
-        test: stylusFiles,
-        exclude: stylusModulesFiles,
-        loader: `null`,
-      })
-      config.loader(`stylusModules`, {
-        test: stylusModulesFiles,
-        loader: moduleLoader,
-      })
-
-      return config
-    }
-
-    default: {
-      return config
-    }
+      break
   }
+
+  setWebpackConfig({
+    module: {
+      rules: configRules,
+    },
+  })
 }
