@@ -130,7 +130,7 @@ function findBabelPackage(directory) {
 module.exports = async function babelConfig(program, stage) {
   const { directory } = program
 
-  let babelrc = findBabelrc(directory) || findBabelPackage(directory)
+  let babelrc = findBabelrc(directory) || findBabelPackage(directory) || {}
 
   // If user doesn't have a custom babelrc, add defaults.
   if (!babelrc) {
@@ -162,7 +162,7 @@ module.exports = async function babelConfig(program, stage) {
   ].forEach(preset => {
     babelrc.presets.push(preset)
   })
-  ;[`add-module-exports`, `transform-object-assign`].forEach(plugin => {
+  ;[`transform-object-assign`].forEach(plugin => {
     babelrc.plugins.push(plugin)
   })
 
@@ -171,7 +171,7 @@ module.exports = async function babelConfig(program, stage) {
     babelrc.plugins.unshift(`react-hot-loader/babel`)
   }
 
-  babelrc.plugins.unshift(require.resolve(`./babel-plugin-extract-graphql`))
+  babelrc.plugins.unshift(require.resolve(`babel-plugin-remove-graphql-queries`))
 
   if (!babelrc.hasOwnProperty(`cacheDirectory`)) {
     babelrc.cacheDirectory = true
