@@ -1,26 +1,21 @@
 import { uniq, some } from "lodash"
 import fs from "fs"
 import path from "path"
-import webpack from "webpack"
 import dotenv from "dotenv"
-import Config from "webpack-configurator"
-import ExtractTextPlugin from "extract-text-webpack-plugin"
 import StaticSiteGeneratorPlugin from "static-site-generator-webpack-plugin"
 import { StatsWriterPlugin } from "webpack-stats-plugin"
 import FriendlyErrorsWebpackPlugin from "friendly-errors-webpack-plugin"
 
-// This isn't working right it seems.
-// import WebpackStableModuleIdAndHash from 'webpack-stable-module-id-and-hash'
-import webpackModifyValidate from "./webpack-modify-validate"
-
 const { store } = require(`../redux`)
+const { actions } = require(`../redux/actions`)
 const debug = require(`debug`)(`gatsby:webpack-config`)
 const WebpackMD5Hash = require(`webpack-md5-hash`)
 const ChunkManifestPlugin = require(`chunk-manifest-webpack-plugin`)
 const GatsbyModulePlugin = require(`../loaders/gatsby-module-loader/plugin`)
-const genBabelConfig = require(`./babel-config`)
 const { withBasePath } = require(`./path`)
-const HashedChunkIdsPlugin = require(`./hashed-chunk-ids-plugin`)
+
+const apiRunnerNode = require(`./api-runner-node`)
+const createConfig = require(`./webpack-utils`)
 
 // Five stages or modes:
 //   1) develop: for `gatsby develop` command, hot reload and CSS injection into page
@@ -36,13 +31,13 @@ module.exports = async (
   webpackPort = 1500,
   pages = []
 ) => {
-  const babelStage = suppliedStage
   const directoryPath = withBasePath(directory)
 
   // We combine develop & develop-html stages for purposes of generating the
   // webpack config.
   const stage = suppliedStage
-  const babelConfig = await genBabelConfig(program, babelStage)
+  const webpackConfig = await createConfig({ stage, program })
+  const { rules, loaders, plugins } = webpackConfig
 
   function processEnv(stage, defaultNodeEnv) {
     debug(`Building env for "${stage}"`)
@@ -71,7 +66,7 @@ module.exports = async (
   }
 
   debug(`Loading webpack config for stage "${stage}"`)
-  function output() {
+  function getOutput() {
     switch (stage) {
       case `develop`:
         return {
@@ -115,7 +110,7 @@ module.exports = async (
     }
   }
 
-  function entry() {
+  function getEntry() {
     switch (stage) {
       case `develop`:
         return {
@@ -148,23 +143,43 @@ module.exports = async (
     }
   }
 
-  function plugins() {
+  function getPlugins() {
+    const env =
+      stage === `develop` || stage === `develop-html`
+        ? `development`
+        : `production`
+
+    let configPlugins = [
+      plugins.moment(),
+
+      // There seems to be a bug in file-loader that assumes this will be set.
+      plugins.loaderOptions({ fileLoader: {} }),
+
+      // Add a few global variables. Set NODE_ENV to production (enables
+      // optimizations for React) and whether prefixing links is enabled
+      // (__PREFIX_PATHS__) and what the link prefix is (__PATH_PREFIX__).
+      plugins.define({
+        "process.env": processEnv(stage, env),
+        __PREFIX_PATHS__: program.prefixPaths,
+        __PATH_PREFIX__: JSON.stringify(store.getState().config.pathPrefix),
+      }),
+
+      plugins.extractText({
+        filename: stage === `build-css` ? `styles.css` : `${stage}.css`,
+      }),
+    ]
+
     switch (stage) {
       case `develop`:
-        return [
-          new webpack.optimize.OccurenceOrderPlugin(),
-          new webpack.HotModuleReplacementPlugin(),
-          new webpack.NoErrorsPlugin(),
-          new webpack.DefinePlugin({
-            "process.env": processEnv(stage, `development`),
-            __PREFIX_PATHS__: program.prefixPaths,
-            __PATH_PREFIX__: JSON.stringify(store.getState().config.pathPrefix),
-          }),
+        configPlugins = configPlugins.concat([
+          plugins.hotModuleReplacement(),
+          plugins.noErrors(),
+
           // Names module ids with their filepath. We use this in development
           // to make it easier to see what modules have hot reloaded, etc. as
           // the numerical IDs aren't useful. In production we use numerical module
           // ids to reduce filesize.
-          new webpack.NamedModulesPlugin(),
+          plugins.namedModules(),
           new FriendlyErrorsWebpackPlugin({
             compilationSuccessInfo: {
               messages: [
@@ -173,55 +188,31 @@ module.exports = async (
               ],
             },
           }),
-        ]
+        ])
+        break
+
       case `develop-html`:
-        return [
-          new StaticSiteGeneratorPlugin(`render-page.js`, pages),
-          new webpack.DefinePlugin({
-            "process.env": processEnv(stage, `development`),
-            __PREFIX_PATHS__: program.prefixPaths,
-            __PATH_PREFIX__: JSON.stringify(store.getState().config.pathPrefix),
-          }),
-          new ExtractTextPlugin(`build-html-styles.css`),
-        ]
-      case `build-css`:
-        return [
-          new webpack.DefinePlugin({
-            "process.env": processEnv(stage, `production`),
-            __PREFIX_PATHS__: program.prefixPaths,
-            __PATH_PREFIX__: JSON.stringify(store.getState().config.pathPrefix),
-          }),
-          new ExtractTextPlugin(`styles.css`, { allChunks: true }),
-        ]
       case `build-html`:
-        return [
+        configPlugins = configPlugins.concat([
           new StaticSiteGeneratorPlugin(`render-page.js`, pages),
-          new webpack.DefinePlugin({
-            "process.env": processEnv(stage, `production`),
-            __PREFIX_PATHS__: program.prefixPaths,
-            __PATH_PREFIX__: JSON.stringify(store.getState().config.pathPrefix),
-          }),
-          new ExtractTextPlugin(`build-html-styles.css`),
-        ]
+        ])
+        break
+
       case `build-javascript`: {
         // Get array of page template component names.
         let components = store
           .getState()
           .pages.map(page => page.componentChunkName)
+
         components = uniq(components)
         components.push(`layout-component---index`)
-        return [
-          // Moment.js includes 100s of KBs of extra localization data by
-          // default in Webpack that most sites don't want. This line disables
-          // loading locale modules. This is a practical solution that requires
-          // the user to opt into importing specific locales.
-          // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-          new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+
+        configPlugins = configPlugins.concat([
           new WebpackMD5Hash(),
-          // new webpack.optimize.DedupePlugin(),
+
           // Extract "commons" chunk from the app entry and all
           // page components.
-          new webpack.optimize.CommonsChunkPlugin({
+          plugins.commonsChunk({
             name: `commons`,
             chunks: [`app`, ...components],
             // The more page components there are, the higher we raise the bar
@@ -259,26 +250,18 @@ module.exports = async (
               ]
               const isFramework = some(
                 vendorModuleList.map(vendor => {
-                  const regex = new RegExp(`\/node_modules\/${vendor}\/.*`, `i`)
+                  const regex = new RegExp(`/node_modules/${vendor}/.*`, `i`)
                   return regex.test(module.resource)
                 })
               )
               return isFramework || count > 3
             },
           }),
-          // Add a few global variables. Set NODE_ENV to production (enables
-          // optimizations for React) and whether prefixing links is enabled
-          // (__PREFIX_PATHS__) and what the link prefix is (__PATH_PREFIX__).
-          new webpack.DefinePlugin({
-            "process.env": processEnv(stage, `production`),
-            __PREFIX_PATHS__: program.prefixPaths,
-            __PATH_PREFIX__: JSON.stringify(store.getState().config.pathPrefix),
-          }),
-          // Extract CSS so it doesn't get added to JS bundles.
-          new ExtractTextPlugin(`build-js-styles.css`),
+
           // Write out mapping between chunk names and their hashed names. We use
           // this to add the needed javascript files to each HTML page.
           new StatsWriterPlugin(),
+
           // Extract the webpack chunk manifest out of commons.js so commons.js
           // doesn't get changed everytime you build. This increases the cache-hit
           // rate for commons.js.
@@ -287,154 +270,68 @@ module.exports = async (
             manifestVariable: `webpackManifest`,
           }),
           // Minify Javascript.
-          new webpack.optimize.UglifyJsPlugin({
-            compress: {
-              screw_ie8: true, // React doesn't support IE8
-              warnings: false,
-            },
-            mangle: {
-              screw_ie8: true,
-            },
-            output: {
-              comments: false,
-              screw_ie8: true,
-            },
-          }),
-          // Ensure module order stays the same. Supposibly fixed in webpack 2.0.
-          new webpack.optimize.OccurenceOrderPlugin(),
+          plugins.uglify(),
           new GatsbyModulePlugin(),
-          // new WebpackStableModuleIdAndHash({ seed: 9, hashSize: 47 }),
-          new webpack.NamedModulesPlugin(),
-          new HashedChunkIdsPlugin(),
-        ]
+          plugins.namedModules(),
+          plugins.namedChunks(),
+        ])
+        break
       }
-      default:
-        throw new Error(`The state requested ${stage} doesn't exist.`)
     }
+
+    return configPlugins
   }
 
-  function resolve() {
-    const { program } = store.getState()
-    return {
-      // Use the program's extension list (generated via the
-      // 'resolvableExtensions' API hook).
-      extensions: [``, ...program.extensions],
-      // Default to using the site's node_modules directory to look for
-      // modules. But also make it possible to install modules within the src
-      // directory if you need to install a specific version of a module for a
-      // part of your site.
-      modulesDirectories: [
-        `node_modules`,
-        directoryPath(`node_modules`),
-        directoryPath(`node_modules`, `gatsby/node_modules`),
-      ],
-    }
-  }
-
-  function devtool() {
+  function getDevtool() {
     switch (stage) {
       case `develop`:
         return `cheap-module-source-map`
-      case `build-html`:
-      case `develop-html`:
-        return false
       case `build-javascript`:
         return `source-map`
+      case `build-html`:
+      case `develop-html`:
       default:
         return false
     }
   }
 
-  function module(config) {
+  function getModule(config) {
+    const browsers = program.browserslist
+    const postcssPlugins = loader => [
+      require(`postcss-import`)({ root: loader.resourcePath }),
+      require(`postcss-cssnext`)({
+        browsers,
+        features: {
+          rem: false, // only needed for <= ie8
+          autoprefixer: false, // handled already
+        },
+      }),
+      require(`postcss-browser-reporter`),
+      require(`postcss-reporter`),
+    ]
+
+    const cssRule = rules.css({ plugins: postcssPlugins })
+    const cssModulesRule = rules.cssModules({
+      plugins: postcssPlugins,
+    })
+
     // Common config for every env.
-    config.loader(`js`, {
-      test: /\.jsx?$/, // Accept either .js or .jsx files.
-      exclude: /(node_modules|bower_components)/,
-      loader: `babel`,
-      query: babelConfig,
-    })
-    config.loader(`json`, {
-      test: /\.json$/,
-      loaders: [`json`],
-    })
-    config.loader(`yaml`, {
-      test: /\.ya?ml/,
-      loaders: [`json`, `yaml`],
-    })
-
-    // "file" loader makes sure those assets end up in the `public` folder.
-    // When you `import` an asset, you get its filename.
-    config.loader(`file-loader`, {
-      test: /\.(ico|eot|otf|webp|ttf|woff(2)?)(\?.*)?$/,
-      loader: `file`,
-      query: {
-        name: `static/[name].[hash:8].[ext]`,
-      },
-    })
-    // "url" loader works just like "file" loader but it also embeds
-    // assets smaller than specified size as data URLs to avoid requests.
-    config.loader(`url-loader`, {
-      test: /\.(svg|jpg|jpeg|png|gif|mp4|webm|wav|mp3|m4a|aac|oga)(\?.*)?$/,
-      loader: `url`,
-      query: {
-        limit: 10000,
-        name: `static/[name].[hash:8].[ext]`,
-      },
-    })
-
-    const cssModulesConf = `css?modules&minimize&importLoaders=1`
-    const cssModulesConfDev = `${cssModulesConf}&sourceMap&localIdentName=[name]---[local]---[hash:base64:5]`
+    // prettier-ignore
+    let configRules = [
+      rules.js(),
+      rules.yaml(),
+      rules.assets(),
+      rules.images(),
+    ]
 
     switch (stage) {
       case `develop`:
-        config.loader(`css`, {
-          test: /\.css$/,
-          exclude: /\.module\.css$/,
-          loaders: [`style`, `css`, `postcss`],
-        })
-
-        // CSS modules
-        config.loader(`cssModules`, {
-          test: /\.module\.css$/,
-          loaders: [`style`, cssModulesConfDev, `postcss`],
-        })
-
-        config.merge({
-          postcss(wp) {
-            return [
-              require(`postcss-import`)({ addDependencyTo: wp }),
-              require(`postcss-cssnext`)({ browsers: program.browserslist }),
-              require(`postcss-browser-reporter`),
-              require(`postcss-reporter`),
-            ]
-          },
-        })
-        return config
+        configRules = configRules.concat([cssRule, cssModulesRule])
+        break
 
       case `build-css`:
-        config.loader(`css`, {
-          test: /\.css$/,
-          exclude: /\.module\.css$/,
-          loader: ExtractTextPlugin.extract([`css?minimize`, `postcss`]),
-        })
-
-        // CSS modules
-        config.loader(`cssModules`, {
-          test: /\.module\.css$/,
-          loader: ExtractTextPlugin.extract(`style`, [
-            cssModulesConf,
-            `postcss`,
-          ]),
-        })
-        config.merge({
-          postcss: [
-            require(`postcss-import`)(),
-            require(`postcss-cssnext`)({
-              browsers: program.browserslist,
-            }),
-          ],
-        })
-        return config
+        configRules = configRules.concat([cssRule, cssModulesRule])
+        break
 
       case `build-html`:
       case `develop-html`:
@@ -442,22 +339,15 @@ module.exports = async (
         // The 'null' loader is used to prevent 'module not found' errors.
         // On the other hand CSS modules loaders are necessary.
 
-        config.loader(`css`, {
-          test: /\.css$/,
-          exclude: /\.module\.css$/,
-          loader: `null`,
-        })
-
-        // CSS modules
-        config.loader(`cssModules`, {
-          test: /\.module\.css$/,
-          loader: ExtractTextPlugin.extract(`style`, [
-            cssModulesConf,
-            `postcss`,
-          ]),
-        })
-
-        return config
+        // prettier-ignore
+        configRules = configRules.concat([
+          {
+            ...cssRule,
+            use: loaders.null,
+          },
+          cssModulesRule,
+        ])
+        break
 
       case `build-javascript`:
         // we don't deal with css at all when building the javascript.  but
@@ -466,31 +356,32 @@ module.exports = async (
         //
         // It's also necessary to process CSS Modules so your JS knows the
         // classNames to use.
+        configRules = configRules.concat([cssRule, cssModulesRule])
+        break
+    }
 
-        config.loader(`css`, {
-          test: /\.css$/,
-          exclude: /\.module\.css$/,
-          // loader: `null`,
-          loader: ExtractTextPlugin.extract([`css`]),
-        })
+    return { rules: configRules }
+  }
 
-        // CSS modules
-        config.loader(`cssModules`, {
-          test: /\.module\.css$/,
-          loader: ExtractTextPlugin.extract(`style`, [
-            cssModulesConf,
-            `postcss`,
-          ]),
-        })
-
-        return config
-
-      default:
-        return config
+  function getResolve() {
+    const { program } = store.getState()
+    return {
+      // Use the program's extension list (generated via the
+      // 'resolvableExtensions' API hook).
+      extensions: [...program.extensions],
+      // Default to using the site's node_modules directory to look for
+      // modules. But also make it possible to install modules within the src
+      // directory if you need to install a specific version of a module for a
+      // part of your site.
+      modules: [
+        `node_modules`,
+        directoryPath(`node_modules`),
+        directoryPath(`node_modules`, `gatsby/node_modules`),
+      ],
     }
   }
 
-  function resolveLoader() {
+  function getResolveLoader() {
     const root = [path.resolve(directory, `node_modules`)]
 
     const userLoaderDirectoryPath = path.resolve(directory, `loaders`)
@@ -506,39 +397,45 @@ module.exports = async (
     }
 
     return {
-      root,
-      modulesDirectories: [path.join(__dirname, `../loaders`), `node_modules`],
+      modules: [...root, path.join(__dirname, `../loaders`), `node_modules`],
     }
   }
 
-  const config = new Config()
-
-  config.merge({
+  const config = {
     // Context is the base directory for resolving the entry option.
     context: directory,
-    node: {
-      __filename: true,
-    },
-    entry: entry(),
-    debug: true,
+    entry: getEntry(),
+    output: getOutput(),
+
+    module: getModule(),
+    plugins: getPlugins(),
+
     // Certain "isomorphic" packages have different entry points for browser
     // and server (see
     // https://github.com/defunctzombie/package-browser-field-spec); setting
     // the target tells webpack which file to include, ie. browser vs main.
     target: stage === `build-html` || stage === `develop-html` ? `node` : `web`,
     profile: stage === `production`,
-    devtool: devtool(),
-    output: output(),
-    resolveLoader: resolveLoader(),
-    plugins: plugins(),
-    resolve: resolve(),
+    devtool: getDevtool(),
+
+    resolveLoader: getResolveLoader(),
+    resolve: getResolve(),
+
+    node: {
+      __filename: true,
+    },
+  }
+
+  store.dispatch(actions.replaceWebpackConfig(config))
+  const getConfig = () => store.getState().webpack
+
+  await apiRunnerNode(`modifyWebpackConfig`, {
+    getConfig,
+    stage,
+    rules,
+    loaders,
+    plugins,
   })
 
-  module(config, stage)
-
-  // Use the suppliedStage again to let plugins distinguish between
-  // server rendering the html.js and the frontend development config.
-  const validatedConfig = await webpackModifyValidate(config, suppliedStage)
-
-  return validatedConfig
+  return getConfig()
 }
