@@ -7,7 +7,7 @@ import {
   TABLET_MEDIA_QUERY,
 } from "typography-breakpoint-constants"
 
-const options = {
+const _options = {
   headerFontFamily: [`Futura PT`, `sans-serif`],
   bodyFontFamily: [`Tex Gyre Schola`, `serif`],
   baseFontSize: `18px`,
@@ -202,11 +202,12 @@ const options = {
   },
 }
 
-const typography = new Typography(options)
+const typography = new Typography(_options)
 
 // Hot reload typography in development.
 if (process.env.NODE_ENV !== `production`) {
   typography.injectStyles()
 }
 
+export const { scale, rhythm, options } = typography
 export default typography
